Drop "wagtail" prefix from contrib.wagtailapi
