+++ wagtail/contrib/api/pagination.py
