+++ wagtail/contrib/api/signal_handlers.py
