+++ wagtail/contrib/api/tests/test_documents.py
@@ -5,19 +5,19 @@ from django.test import TestCase
 from django.test.utils import override_settings
 from django.core.urlresolvers import reverse
 
-from wagtail.wagtailimages.models import get_image_model
+from wagtail.wagtaildocs.models import Document
 
-from wagtail.contrib.wagtailapi import signal_handlers
+from .. import signal_handlers
 
 
-class TestImageListing(TestCase):
+class TestDocumentListing(TestCase):
     fixtures = ['demosite.json']
 
     def get_response(self, **params):
-        return self.client.get(reverse('wagtailapi_v1:images:listing'), params)
+        return self.client.get(reverse('wagtailapi_v1:documents:listing'), params)
 
-    def get_image_id_list(self, content):
-        return [page['id'] for page in content['images']]
+    def get_document_id_list(self, content):
+        return [page['id'] for page in content['documents']]
 
 
     # BASIC TESTS
@@ -38,23 +38,23 @@ class TestImageListing(TestCase):
         # Check that the total count is there and correct
         self.assertIn('total_count', content['meta'])
         self.assertIsInstance(content['meta']['total_count'], int)
-        self.assertEqual(content['meta']['total_count'], get_image_model().objects.count())
+        self.assertEqual(content['meta']['total_count'], Document.objects.count())
 
-        # Check that the images section is there
-        self.assertIn('images', content)
-        self.assertIsInstance(content['images'], list)
+        # Check that the documents section is there
+        self.assertIn('documents', content)
+        self.assertIsInstance(content['documents'], list)
 
-        # Check that each image has a meta section with type and detail_url attributes
-        for image in content['images']:
-            self.assertIn('meta', image)
-            self.assertIsInstance(image['meta'], dict)
-            self.assertEqual(set(image['meta'].keys()), {'type', 'detail_url'})
+        # Check that each document has a meta section with type and detail_url attributes
+        for document in content['documents']:
+            self.assertIn('meta', document)
+            self.assertIsInstance(document['meta'], dict)
+            self.assertEqual(set(document['meta'].keys()), {'type', 'detail_url'})
 
-            # Type should always be wagtailimages.Image
-            self.assertEqual(image['meta']['type'], 'wagtailimages.Image')
+            # Type should always be wagtaildocs.Document
+            self.assertEqual(document['meta']['type'], 'wagtaildocs.Document')
 
-            # Check detail url
-            self.assertEqual(image['meta']['detail_url'], 'http://localhost/api/v1/images/%d/' % image['id'])
+            # Check detail_url
+            self.assertEqual(document['meta']['detail_url'], 'http://localhost/api/v1/documents/%d/' % document['id'])
 
 
     # EXTRA FIELDS
@@ -63,23 +63,22 @@ class TestImageListing(TestCase):
         response = self.get_response()
         content = json.loads(response.content.decode('UTF-8'))
 
-        for image in content['images']:
-            self.assertEqual(set(image.keys()), {'id', 'meta', 'title'})
+        for document in content['documents']:
+            self.assertEqual(set(document.keys()), {'id', 'meta', 'title'})
 
     def test_extra_fields(self):
-        response = self.get_response(fields='title,width,height')
+        response = self.get_response(fields='title,tags')
         content = json.loads(response.content.decode('UTF-8'))
 
-        for image in content['images']:
-            self.assertEqual(set(image.keys()), {'id', 'meta', 'title', 'width', 'height'})
+        for document in content['documents']:
+            self.assertEqual(set(document.keys()), {'id', 'meta', 'title', 'tags'})
 
     def test_extra_fields_tags(self):
         response = self.get_response(fields='tags')
         content = json.loads(response.content.decode('UTF-8'))
 
-        for image in content['images']:
-            self.assertEqual(set(image.keys()), {'id', 'meta', 'tags'})
-            self.assertIsInstance(image['tags'], list)
+        for document in content['documents']:
+            self.assertIsInstance(document['tags'], list)
 
     def test_extra_fields_which_are_not_in_api_fields_gives_error(self):
         response = self.get_response(fields='uploaded_by_user')
@@ -102,24 +101,24 @@ class TestImageListing(TestCase):
         response = self.get_response(title='James Joyce')
         content = json.loads(response.content.decode('UTF-8'))
 
-        image_id_list = self.get_image_id_list(content)
-        self.assertEqual(image_id_list, [5])
+        document_id_list = self.get_document_id_list(content)
+        self.assertEqual(document_id_list, [2])
 
     def test_filtering_on_id(self):
         response = self.get_response(id=10)
         content = json.loads(response.content.decode('UTF-8'))
 
-        image_id_list = self.get_image_id_list(content)
-        self.assertEqual(image_id_list, [10])
+        document_id_list = self.get_document_id_list(content)
+        self.assertEqual(document_id_list, [10])
 
     def test_filtering_tags(self):
-        get_image_model().objects.get(id=6).tags.add('test')
+        Document.objects.get(id=3).tags.add('test')
 
         response = self.get_response(tags='test')
         content = json.loads(response.content.decode('UTF-8'))
 
-        image_id_list = self.get_image_id_list(content)
-        self.assertEqual(image_id_list, [6])
+        document_id_list = self.get_document_id_list(content)
+        self.assertEqual(document_id_list, [3])
 
     def test_filtering_unknown_field_gives_error(self):
         response = self.get_response(not_a_field='abc')
@@ -135,26 +134,26 @@ class TestImageListing(TestCase):
         response = self.get_response(order='title')
         content = json.loads(response.content.decode('UTF-8'))
 
-        image_id_list = self.get_image_id_list(content)
-        self.assertEqual(image_id_list, [6, 15, 13, 5, 10, 11, 8, 7, 4, 14, 12, 9])
+        document_id_list = self.get_document_id_list(content)
+        self.assertEqual(document_id_list, [3, 12, 10, 2, 7, 8, 5, 4, 1, 11, 9, 6])
 
     def test_ordering_by_title_backwards(self):
         response = self.get_response(order='-title')
         content = json.loads(response.content.decode('UTF-8'))
 
-        image_id_list = self.get_image_id_list(content)
-        self.assertEqual(image_id_list, [9, 12, 14, 4, 7, 8, 11, 10, 5, 13, 15, 6])
+        document_id_list = self.get_document_id_list(content)
+        self.assertEqual(document_id_list, [6, 9, 11, 1, 4, 5, 8, 7, 2, 10, 12, 3])
 
     def test_ordering_by_random(self):
         response_1 = self.get_response(order='random')
         content_1 = json.loads(response_1.content.decode('UTF-8'))
-        image_id_list_1 = self.get_image_id_list(content_1)
+        document_id_list_1 = self.get_document_id_list(content_1)
 
         response_2 = self.get_response(order='random')
         content_2 = json.loads(response_2.content.decode('UTF-8'))
-        image_id_list_2 = self.get_image_id_list(content_2)
+        document_id_list_2 = self.get_document_id_list(content_2)
 
-        self.assertNotEqual(image_id_list_1, image_id_list_2)
+        self.assertNotEqual(document_id_list_1, document_id_list_2)
 
     def test_ordering_by_random_backwards_gives_error(self):
         response = self.get_response(order='-random')
@@ -184,14 +183,14 @@ class TestImageListing(TestCase):
         response = self.get_response(limit=2)
         content = json.loads(response.content.decode('UTF-8'))
 
-        self.assertEqual(len(content['images']), 2)
+        self.assertEqual(len(content['documents']), 2)
 
     def test_limit_total_count(self):
         response = self.get_response(limit=2)
         content = json.loads(response.content.decode('UTF-8'))
 
         # The total count must not be affected by "limit"
-        self.assertEqual(content['meta']['total_count'], get_image_model().objects.count())
+        self.assertEqual(content['meta']['total_count'], Document.objects.count())
 
     def test_limit_not_integer_gives_error(self):
         response = self.get_response(limit='abc')
@@ -222,29 +221,29 @@ class TestImageListing(TestCase):
         response = self.get_response()
         content = json.loads(response.content.decode('UTF-8'))
 
-        self.assertEqual(len(content['images']), 2)
+        self.assertEqual(len(content['documents']), 2)
 
 
     # OFFSET
 
-    def test_offset_10_usually_appears_7th_in_list(self):
+    def test_offset_5_usually_appears_5th_in_list(self):
         response = self.get_response()
         content = json.loads(response.content.decode('UTF-8'))
-        image_id_list = self.get_image_id_list(content)
-        self.assertEqual(image_id_list.index(10), 6)
+        document_id_list = self.get_document_id_list(content)
+        self.assertEqual(document_id_list.index(5), 4)
 
-    def test_offset_10_moves_after_offset(self):
+    def test_offset_5_moves_after_offset(self):
         response = self.get_response(offset=4)
         content = json.loads(response.content.decode('UTF-8'))
-        image_id_list = self.get_image_id_list(content)
-        self.assertEqual(image_id_list.index(10), 2)
+        document_id_list = self.get_document_id_list(content)
+        self.assertEqual(document_id_list.index(5), 0)
 
     def test_offset_total_count(self):
         response = self.get_response(offset=10)
         content = json.loads(response.content.decode('UTF-8'))
 
         # The total count must not be affected by "offset"
-        self.assertEqual(content['meta']['total_count'], get_image_model().objects.count())
+        self.assertEqual(content['meta']['total_count'], Document.objects.count())
 
     def test_offset_not_integer_gives_error(self):
         response = self.get_response(offset='abc')
@@ -260,9 +259,9 @@ class TestImageListing(TestCase):
         response = self.get_response(search='james')
         content = json.loads(response.content.decode('UTF-8'))
 
-        image_id_list = self.get_image_id_list(content)
+        document_id_list = self.get_document_id_list(content)
 
-        self.assertEqual(set(image_id_list), set([5]))
+        self.assertEqual(set(document_id_list), set([2]))
 
     def test_search_when_ordering_gives_error(self):
         response = self.get_response(search='james', order='title')
@@ -287,15 +286,14 @@ class TestImageListing(TestCase):
         self.assertEqual(content, {'message': "filtering by tag with a search query is not supported"})
 
 
-class TestImageDetail(TestCase):
+class TestDocumentDetail(TestCase):
     fixtures = ['demosite.json']
 
     def get_response(self, image_id, **params):
-        return self.client.get(reverse('wagtailapi_v1:images:detail', args=(image_id, )), params)
-
+        return self.client.get(reverse('wagtailapi_v1:documents:detail', args=(image_id, )), params)
 
     def test_basic(self):
-        response = self.get_response(5)
+        response = self.get_response(1)
 
         self.assertEqual(response.status_code, 200)
         self.assertEqual(response['Content-type'], 'application/json')
@@ -305,7 +303,7 @@ class TestImageDetail(TestCase):
 
         # Check the id field
         self.assertIn('id', content)
-        self.assertEqual(content['id'], 5)
+        self.assertEqual(content['id'], 1)
 
         # Check that the meta section is there
         self.assertIn('meta', content)
@@ -313,36 +311,43 @@ class TestImageDetail(TestCase):
 
         # Check the meta type
         self.assertIn('type', content['meta'])
-        self.assertEqual(content['meta']['type'], 'wagtailimages.Image')
+        self.assertEqual(content['meta']['type'], 'wagtaildocs.Document')
 
         # Check the meta detail_url
         self.assertIn('detail_url', content['meta'])
-        self.assertEqual(content['meta']['detail_url'], 'http://localhost/api/v1/images/5/')
+        self.assertEqual(content['meta']['detail_url'], 'http://localhost/api/v1/documents/1/')
+
+        # Check the meta download_url
+        self.assertIn('download_url', content['meta'])
+        self.assertEqual(content['meta']['download_url'], 'http://localhost/documents/1/wagtail_by_markyharky.jpg')
 
         # Check the title field
         self.assertIn('title', content)
-        self.assertEqual(content['title'], "James Joyce")
-
-        # Check the width and height fields
-        self.assertIn('width', content)
-        self.assertIn('height', content)
-        self.assertEqual(content['width'], 500)
-        self.assertEqual(content['height'], 392)
+        self.assertEqual(content['title'], "Wagtail by mark Harkin")
 
         # Check the tags field
         self.assertIn('tags', content)
         self.assertEqual(content['tags'], [])
 
     def test_tags(self):
-        image = get_image_model().objects.get(id=5)
-        image.tags.add('hello')
-        image.tags.add('world')
+        Document.objects.get(id=1).tags.add('hello')
+        Document.objects.get(id=1).tags.add('world')
 
-        response = self.get_response(5)
+        response = self.get_response(1)
         content = json.loads(response.content.decode('UTF-8'))
 
         self.assertIn('tags', content)
         self.assertEqual(content['tags'], ['hello', 'world'])
+
+    @override_settings(WAGTAILAPI_BASE_URL='http://api.example.com/')
+    def test_download_url_with_custom_base_url(self):
+        response = self.get_response(1)
+        content = json.loads(response.content.decode('UTF-8'))
+
+        self.assertIn('download_url', content['meta'])
+        self.assertEqual(
+            content['meta']['download_url'], 'http://api.example.com/documents/1/wagtail_by_markyharky.jpg'
+        )
 
 
 @override_settings(
@@ -355,25 +360,25 @@ class TestImageDetail(TestCase):
     WAGTAILAPI_BASE_URL='http://api.example.com',
 )
 @mock.patch('wagtail.contrib.wagtailfrontendcache.backends.HTTPBackend.purge')
-class TestImageCacheInvalidation(TestCase):
+class TestDocumentCacheInvalidation(TestCase):
     fixtures = ['demosite.json']
 
     @classmethod
     def setUpClass(cls):
-        super(TestImageCacheInvalidation, cls).setUpClass()
+        super(TestDocumentCacheInvalidation, cls).setUpClass()
         signal_handlers.register_signal_handlers()
 
     @classmethod
     def tearDownClass(cls):
-        super(TestImageCacheInvalidation, cls).tearDownClass()
+        super(TestDocumentCacheInvalidation, cls).tearDownClass()
         signal_handlers.unregister_signal_handlers()
 
-    def test_resave_image_purges(self, purge):
-        get_image_model().objects.get(id=5).save()
+    def test_resave_document_purges(self, purge):
+        Document.objects.get(id=5).save()
 
-        purge.assert_any_call('http://api.example.com/api/v1/images/5/')
+        purge.assert_any_call('http://api.example.com/api/v1/documents/5/')
 
-    def test_delete_image_purges(self, purge):
-        get_image_model().objects.get(id=5).delete()
+    def test_delete_document_purges(self, purge):
+        Document.objects.get(id=5).delete()
 
-        purge.assert_any_call('http://api.example.com/api/v1/images/5/')
+        purge.assert_any_call('http://api.example.com/api/v1/documents/5/')
